fix(perf): wire MemoryEstimator into MaterialEditorQML test targets + CR feedback
CI failure on #494: MaterialEditorQML_perf_test / _qml_test linked
against MCPServer / CLIPipeline / MeshInfoOverlay sources but the new
MemoryEstimator.cpp was missing from tests/CMakeLists.txt's source
list, producing undefined references on the GCC link stage. Added it
alongside TextureChannelPacker.cpp / NormalMapGenerator.cpp.

CodeRabbit feedback addressed in the same commit (keep the diff small):

- MCP get_memory_usage now returns the SceneMemoryReport as a structured
  `memory` field in the result envelope rather than appending a JSON
  string to the human summary. Machine consumers no longer have to
  text-parse the response.
- MemoryEstimator::estimateScene now iterates node->numAttachedObjects()
  with an unsigned counter, matching the Ogre API and dropping the
  static_cast<int> noise.
- qtmesh memory --budget error message clarifies that omitting the flag
  means unlimited (was ambiguous on `--budget 0`).

Also adds cloud-rules integration the user asked for: when --budget is
not passed and a QtMesh Cloud token is available (--token / QTMESH_TOKEN
/ QTMESH_CLOUD_TOKEN), `qtmesh memory` GETs /v1/ingest/rules and reads
rules.memory_budget_mb from the project config. --no-cloud opts out.
The selected source is recorded in Sentry breadcrumbs and in --json
output (budgetSource: "cli" | "cloud:<source>").

+++ src/MemoryEstimator.cpp
@@ -166,7 +166,7 @@ SceneMemoryReport MemoryEstimator::estimateScene(quint64 budgetBytes)
     QSet<QString> seenMeshes;
     for (Ogre::SceneNode* node : mgr->getSceneNodes()) {
         if (!node) continue;
-        for (int i = 0; i < static_cast<int>(node->numAttachedObjects()); ++i) {
+        for (unsigned int i = 0; i < node->numAttachedObjects(); ++i) {
             Ogre::MovableObject* obj = node->getAttachedObject(i);
             if (!obj || obj->getMovableType() != "Entity") continue;
